refactor(apifrontend): Phase 6 Wave 1 - decompose 5 top-complexity functions, split 3 oversized files
Reduces the highest-cyclomatic-complexity APIFrontend functions left open
after Phase 4 and splits its three largest files, closing out Wave 1 of the
Phase 6+ complexity/file-size burndown (GO-ANTIPATTERN-AUDIT-2026-07-01).

Function decomposition (cyclomatic complexity before -> after):
- buildEventData 38->2 (audit/store_adapter.go): registry pattern, reusing
  the DD-AUDIT-008 approach already applied to KubernautAgent
- RegisterTools 37->4 (handler/mcp_bridge.go): split into domain-specific
  register<Domain>Tools helpers behind a thin dispatcher
- newPhaseGuard 43->1 (agent/phase_guard.go): extracted closure bodies into
  standalone named functions
- handleSubscribe 40->7 (handler/status_handler.go): subscribeStreamState +
  subscribeCleanup + extracted event-loop handlers; characterization via the
  existing status_subscribe_test.go IT suite (12 specs, re-verified pre/post)
- HandleInvestigationMCPWithRegistry 67->10 (tools/ka_investigate_mcp.go):
  extracted RR resolution, KA session start, and blocking/non-blocking
  bridge handoff into named helpers

File splits (same-package, zero behavior change):
- tools/ka_investigate_mcp.go 1111->685 lines (+ ka_investigate_bridge.go,
  ka_investigate_registry.go)
- tools/crd_tools.go 1144->606 lines (+ crd_tools_watch.go,
  crd_tools_session.go)
- cmd/apifrontend/main.go 1701->609 lines (+ backend_deps.go,
  auth_wiring.go, session_infra.go, mcp_a2a_handlers.go)

Fixed 4 argument-limit violations introduced by the extraction itself
(subscribeLoopCtx / blockingInvestigationParams structs) so every touched
function stays within AGENTS.md's 7-argument rule.

Verification: go build/vet ./... (repo-wide), golangci-lint run 0 issues,
full pkg/apifrontend + cmd/apifrontend unit suites green.
